slave clean up
improvements on the slave system. it's player dependent and unglobalized. slaves are controlled by two different kinds of rules: a draw rule and a logic rule. this is necessary because some slaves need special appearances like Myon and drawing is only possible in stage_draw(). this also makes them configurable enough for different shot types and stuff.