Fixed a replay desync related to the stage 6 boss
There was a piece of code in the Scythe drawing function that altered the game state. I hope I don't have to explain why that is a vector for bugs in general, but in particular, this always breaks replays in frameskip mode.

I've also moved the BGM handling from stage_draw to stage_logic for similar reasons.