Removed TAISEI_WIN32_CONSOLE. Details below.
Looks like some sdl2main magic is overriding this, so we don't get a console window anyway.
Still, I made the effect of that option the default: the WIN32 flag is no longer passed to add_executable()
This means -mwindows won't be passed to the compiler when linking. That option is effectively useless in our case, and some compilers/linkers do not support it.
For example wclang, a clang wrapper that can be used as a drop-in replacement for mingw64's gcc, will try to fallback to gcc for linking when it encounters that option.