CMake cleanup
Use pkg-config to find the SDL2-related packages. Seems to work better
for cross-compilation environments, at least for me.

Removed the freetype dependency. It's for SDL2_ttf, we don't directly
use it.

Zlib is a dependency of libpng, but I've let it remain an explicit
dependency for now, because my hopefully-soon-to-be-merged PR uses a
function from it.