change version scheme again, hopefully for the last time
{required}, [optional]

The version format for git tags is as follows:

    v{major}[.minor[.patch]]

All of the components must be numeric, i.e. our former hotfix release
v1.0a is not valid and should have been v1.0.1

Ommited optional compoments are assumed to be 0.

Internally, the version string is parsed as follows:

    [v]{major}[.minor[.patch]][-tweak[extrainfo]]

Tweak must be numeric, and represents the ammount of commits since the
last release.

Extrainfo is currently ignored and can be anything. Note that a version
string obtained from "git describe" would contain a commit hash there,
which we may or may not use in the future. Keep this in mind if you're
overriding the version string with -DTAISEI_VERSION_OVERRIDE

The leading v is stripped if present.

Also, the version string is currently not completely validated, but you
will probably get compilation errors in version.c later if it couldn't
be parsed correctly.