improve youmuA controls
most importantly this makes myon actually possible to position with a
gamepad in free-axes mode. this also removes a bunch of hacky movement
state variables maintained by the generic player code just so that
youmuA's very keyboard-oriented input code could use them. the new
controls should be just as easily usable on a gamepad as they are on a
keyboard, and be more responsive on both.

of course, it breaks existing youmuA replays.