README: remove the BGM section (rationale below)
As work on Taisei's soundtrack is currently in progress and bgm.conf
is deprecated and will soon be removed, this section is becoming
increasingly less relevant. I don't think it should be on the main
README. If anyone wants to maintain documentation on customizing and/or
overriding the game's resources, it should be done on the Wiki. We can
later simply link to such Wiki pages from the README.

The gamepad section is also a good candidate for the Wiki, but I'm
letting it stay here for now, as it's at least accurate and technically
relevant.