leanify all possible resources (gfx)
This reduces the data zip size by about 8%, and also gets rid of PNG
warnings about incorrect color profile on some of textures (because all
unnecessary metadata is removed).

Note: I am well aware that this commit bloats the repository, however,
it is bloated already with ancient, much larger blobs that have been
removed from the tree, but remain in the history (e.g. what is now in
rawmedia). We can (and probably should) eventually get rid of the
old useless blobs by rewriting git history and force pushing. There are
known ways to do exactly this and tools exist specifically for this
purpose. I have tested one of them on a local Taisei clone.

If we end up going through with it, this change is also a net benefit
for the repository size.