confusing and needlessly complicated shaders refactoring
Now using explicit locations everywhere to make SPIRV-cross happy and
arcane macros to make people who know GLSL angry and/or perplexed.

Interface matching is supposed to rely on explicit locations in MODERN
OPENGL (tm) anyway.

Unfortunately, the glslc validation process will not catch every
mistake (or whatever your opengl implementation happens to dislike).

It's recommended to run taisei with TAISEI_PRELOAD_SHADERS=1 to catch
errors in the shaders faster. It will attempt to load them all very
early, and crash if something fails. This is not a substitute for the
glslc tests, however: both have to pass.