Encode and write screenshots on a background thread
This reduces in-game stuttering when a screenshot is taken, especially
if multiple are taken during a short time frame. The pixel transfer from
the renderer to client memory is still synchronous, so some spikes are
to be expected; however, png encoding is usually much slower.

Also, TODO: convert this code into a generic multi-threaded task
manager/worker pool (similar to python's futures), and make the
asynchronous resource loading code suck less using it. Another useful
application is asynchronous logging, but that probably should be turned
off in debug builds by default.