resources: The Great Ref-fucktoring
i'm too tired to explain this but i'll try

Basically it's a way to access and manage lifetime of resources that
makes slightly more sense than... whatever there was before it.
Resources are accessed through explicit reference types, those with no
more references are pending unload, etc. etc. etc. Type-safety is
sacrificed (although it can be restored to an extent with some evil
preprocessor hackery), it's clunky to use from C, and a lot of code
still uses the "old style" through a compatibility layer, but script
bindings are expected to solve these problems.

tagged pointers are used because they're cool

There are also a few other random/tangentally related improvements.