Add option to convert texture into linear color space on load
If "linearize = true" is specified in the .tex file, colors will be
converted from the sRGB color space into the linear RGB color space on
load. Ideally this should be the default, and we should be doing all
color manipulations in linear space, but there's too many things to fix
to make that a reality.

Additionally, "multiply_alpha = false" to prevent colors from being
multiplied by the alpha channel on load.

The texture_post_load shader will no longer be applied when the
preprocessing operation is a no-op.