Fix use-after-free in resource unloading
There is an edge case when a loading resource that has emitted an
asyncload event gets unloaded before that event is handled. In this case
the event is left in the queue with a dangling pointer to a now invalid
resource, causing a use-after-free bug later. This patch makes sure any
such events are removed from the queue before the resource is freed.

Note that this does not leak memory. Unloading can only happen on the
main thread, and the unload function will first finalize any partial
loads. The removed event would have been a no-op.