Asynchronous logging (#260)
* Implement asynchronous logging

Put all log messages into a queue. A background thread will format and
dispatch the messages to loggers in a first-in-first-out fashion. This
is mostly useful for windows/wine, or in cases *very* spammy logging.
This adds some malloc overhead, however.

Can be disabled with TAISEI_LOG_ASYNC=0

* log: unfuck memory management

* log: add TAISEI_LOG_ASYNC_FAST_SHUTDOWN env variable

If true, don't wait for the logging queue to process every message
before shutting down.

* strbuf: fix off-by-one error

* log: sync queue before crashing on assertion failure

* docs/ENVIRON: update logging docs