vfs: transparent decompression of *.zst files
A zstd-compressed file "foobar.foo.zst" will appear in the filesystem as
a normal uncompressed file "foobar.foo". The raw compressed data is
still available under the original .zst filename.

If both "foobar.foo" and "foobar.foo.zst" exist, the former is
preferred and won't be shadowed.

This is implemented as a wrapper layer, much like readonly_wrapper.
It is currently applied on top of plain directories that contribute to
/res, but not to zipped packages. Zip files support zstd as a
compression method, so we'll probably rely on that instead.