coroutine: add a "sized pointer" type for interface arg structs
A sized pointer struct simply contains a pointer to an args struct of
the interface's corresponding type, and a size_t field with the size of
that struct. In case of interface inheritance, the sized pointer can be
"cast" to the base type by accessing the union member "base", enabling a
form of polymorphism.

A sized pointer can be constructed like so:

// alias for convenience
typedef TASK_IFACE_ARGS_SIZED_PTR_TYPE(MyInterface) MyInterfaceSptr;

MyInterfaceSptr sptr = TASK_IFACE_SARGS(MyInterface,
        .foo = bar,
        .spam = eggs
);