coroutine,stage: remove global sched; inherit sched from calling task
This allows having multiple task schedulers without having to switch the
global "target" scheduler for INVOKE_ macros.

An important side effect of this change is that it's not possible to use
the regular INVOKE_ macros from a non-coroutine context anymore. A
series of complimentary SCHED_INVOKE_ macros was added that behave
identically, but allow explicitly specifying a scheduler as the first
argument. The stage loop has been slightly refactored to accomodate this
behavior.