ci: Bump macOS builds to 15.x, default to ARM64.
GHA macOS runners are now ARM64-default, and this required a few changes to
make it work.

We're almost at the release of the M4 processor, which means the Mac
platform is now ~4 years into the move to ARM64, which means we need to
start assuming this as the default.