update zh_CN.po (#410)
Hi, I've translated a portion of the content. (Actually, my local
progress is a bit ahead, but I still haven’t managed to build a working
binary for testing, so I’m hesitant to upload it. Besides, with school
starting soon, time is tight. I’m afraid it might be difficult to
complete the localization before 2026.)